(refactor): separate browser ready builds
