+++ extensions/react-widget/src/index.ts
@@ -1,2 +1,3 @@
+//exports methods for React
 export {SearchBar} from "./components/SearchBar"
 export { DocsGPTWidget } from "./components/DocsGPTWidget";
